Added open-below and open-right shortcuts
Also fixed a bug when switching direction of a container containing a container.
This was due to an accidental copy of a ClientData which only works when it contains a view
(since the copy will have the same pointer and refer to the same view).
To avoid this in the future, made Container non-copiable and changed code to accomodate.