Added `getBool` to `conf::Configuration`, added "enable enter on hover" setting
The new getBool is a simple wrapper to normalize what values we accept for settings until we get an albinos API update.
I updated the "tiling" parameter to "enable default floating" (so it's positive, not a disable).
I added the "enable enter on hover" setting (which I personnaly prefer to have on).

Closes #137